Refactoring code around - working again
