+++ src/assessment_methods/index.jl
@@ -1,3 +1,19 @@
+using
+    Statistics,
+    Proj,
+    LibGEOS,
+    GeometryBasics,
+    CoordinateTransformations,
+    Rasters,
+    StaticArrays,
+    NearestNeighbors
+
+import ArchGDAL as AG
+import GeoInterface as GI
+import GeoInterface.Wrappers as GIWrap
+import GeometryOps as GO
+import GeoDataFrames as GDF
+
 include("apply_criteria.jl")
 include("best_fit_polygons.jl")
 include("common_functions.jl")
