+++ src/utility/types.jl
@@ -1,0 +1,3 @@
+"""Utility types"""
+
+const OptionalValue{T} = Union{T,Nothing};
